requirements.txt
fastapi==0.115.1
uvicorn==0.35.0
sqlalchemy==2.0.41
jinja2==3.1.5
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose==3.5.0
bcrypt==4.0.1
